Rename funding states
