+++ packages/wallet/src/redux/states/channels/funding/sharedState.ts
